+++ Header/get_next_line.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ajabri <ajabri@student.42.fr>              +#+  +:+       +#+        */
+/*   By: kali <kali@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 17:48:29 by ajabri            #+#    #+#             */
-/*   Updated: 2023/12/21 13:19:41 by ajabri           ###   ########.fr       */
+/*   Updated: 2024/10/12 12:36:26 by kali             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ char	*get_next_line(int fd);
 char	*ft_strchr(char *s, int c);
 char	*ft_strdup(char *s1);
 char	*ft_strjoin(char *s1, char *s2);
-size_t	ft_strlen(char *c);
+int	    ft_strlen(char *c);
 void	*ft_memmove(void *dst, void *src, size_t n);
 //___________________________________________________
